move product file to point of use
